Fix parameters parsing when there is a note or warning inside
